Siistitty Navigationia eli vanhaa ApplicationRoutesia.
Nyt AuthStack ja MainStack sekä Drawer ovat kaikki omia komponenttejaan joita käytetään Navigationissa.
