+++ HuiputinPaivakirja/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import { ThemeProvider } from './src/theme/CustomTheme'; // Import ThemeProvider
-import Navigation from './src/Navigation';
+import { ThemeProvider } from './src/theme/CustomTheme';
+import Navigation from './src/navigation/Navigation';
 
 export default function App() {
   return (
